fix: verify URL when viewing dashboards
